+++ src/main/kotlin/org/example/restaurantbackend/dto/response/ReservationResponse.kt
@@ -5,7 +5,7 @@ import java.time.LocalDateTime
 data class ReservationResponse(
     val id: Long,
     val tableId: Long,
-    val restaurantId: Int,
+    val restaurantId: Long,
     val startTime: LocalDateTime,
     val endTime: LocalDateTime,
     val guests: Int,
