+++ src/main/kotlin/org/example/restaurantbackend/repository/FavoriteRepository.kt
@@ -5,7 +5,7 @@ import org.example.restaurantbackend.entity.UserEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface FavoriteRepository : JpaRepository<FavoriteEntity, Long> {
-    fun existsByUserAndDishId(user: UserEntity, dishId: Int): Boolean
-    fun findByUserAndDishId(user: UserEntity, dishId: Int): FavoriteEntity?
+    fun existsByUserAndDishId(user: UserEntity, dishId: Long): Boolean
+    fun findByUserAndDishId(user: UserEntity, dishId: Long): FavoriteEntity?
     fun findAllByUser(user: UserEntity): List<FavoriteEntity>
 }
