+++ src/main/kotlin/org/example/restaurantbackend/repository/NewsRepository.kt
@@ -4,5 +4,5 @@ import org.example.restaurantbackend.entity.NewsEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface NewsRepository : JpaRepository<NewsEntity, Long> {
-    fun findByRestaurantId(restaurantId: Int): List<NewsEntity>
+    fun findByRestaurantId(restaurantId: Long): List<NewsEntity>
 }
